add reset counter

+++ ROS-Driver-Update/roboteq_motor_controller_driver/include/roboteq_motor_controller_driver/roboteq_motor_controller_driver_node.h
@@ -45,10 +45,10 @@ public:
 private:
 
 	int frequency;
-	float wheelbase;
-	float radius;
-	float gearRatio;
-	float maxRPM;
+	double wheelbase;
+	double radius;
+	double gearRatio;
+	double maxRPM;
 	int baud_rate;
 	int32_t baud;
 	std::string port;
